# simplify EstimateBits a bit