--- a/lib/jxl/modular/encoding/enc_ma.cc
+++ b/lib/jxl/modular/encoding/enc_ma.cc
@@ -37,33 +37,23 @@ const HWY_FULL(float) df;
 const HWY_FULL(int32_t) di;
 size_t Padded(size_t x) { return RoundUpTo(x, Lanes(df)); }
 
-float EstimateBits(const int32_t *counts, int32_t *rounded_counts,
-                   size_t num_symbols) {
-  // Try to approximate the effect of rounding up nonzero probabilities.
+// Compute entropy of the histogram, taking into account the minimum probability
+// for symbols with non-zero counts.
+float EstimateBits(const int32_t *counts, size_t num_symbols) {
   int32_t total = std::accumulate(counts, counts + num_symbols, 0);
-  const auto min = Set(di, (total + ANS_TAB_SIZE - 1) >> ANS_LOG_TAB_SIZE);
-  const auto zero_i = Zero(di);
-  for (size_t i = 0; i < num_symbols; i += Lanes(df)) {
-    auto counts_v = LoadU(di, &counts[i]);
-    counts_v = IfThenElse(Eq(counts_v, zero_i), zero_i,
-                          IfThenElse(Lt(counts_v, min), min, counts_v));
-    StoreU(counts_v, di, &rounded_counts[i]);
-  }
-  // Compute entropy of the "rounded" probabilities.
   const auto zero = Zero(df);
-  const size_t total_scalar =
-      std::accumulate(rounded_counts, rounded_counts + num_symbols, 0);
-  const auto inv_total = Set(df, 1.0f / total_scalar);
+  const auto minprob = Set(df, 1.0f / ANS_TAB_SIZE);
+  const auto inv_total = Set(df, 1.0f / total);
   auto bits_lanes = Zero(df);
-  auto total_v = Set(di, total_scalar);
+  auto total_v = Set(di, total);
   for (size_t i = 0; i < num_symbols; i += Lanes(df)) {
-    const auto counts_v = ConvertTo(df, LoadU(di, &counts[i]));
-    const auto round_counts_v = LoadU(di, &rounded_counts[i]);
-    const auto probs = Mul(ConvertTo(df, round_counts_v), inv_total);
-    const auto nbps = IfThenElse(Eq(round_counts_v, total_v), BitCast(di, zero),
-                                 BitCast(di, FastLog2f(df, probs)));
-    bits_lanes = Sub(bits_lanes, IfThenElse(Eq(counts_v, zero), zero,
-                                            Mul(counts_v, BitCast(df, nbps))));
+    const auto counts_iv = LoadU(di, &counts[i]);
+    const auto counts_fv = ConvertTo(df, counts_iv);
+    const auto probs = Mul(counts_fv, inv_total);
+    const auto mprobs = Max(probs, minprob);
+    const auto nbps = IfThenElse(Eq(counts_iv, total_v), BitCast(di, zero),
+                                 BitCast(di, FastLog2f(df, mprobs)));
+    bits_lanes = Sub(bits_lanes, Mul(counts_fv, BitCast(df, nbps)));
   }
   return GetLane(SumOfLanes(df, bits_lanes));
 }
@@ -225,7 +215,6 @@ void FindBestSplit(TreeSamples &tree_samples, float threshold,
       }
     }
     max_symbols = Padded(max_symbols);
-    std::vector<int32_t> rounded_counts(max_symbols);
     std::vector<int32_t> counts(max_symbols * num_predictors);
     std::vector<uint32_t> tot_extra_bits(num_predictors);
     for (size_t pred = 0; pred < num_predictors; pred++) {
@@ -240,9 +229,9 @@ void FindBestSplit(TreeSamples &tree_samples, float threshold,
     float base_bits;
     {
       size_t pred = tree_samples.PredictorIndex((*tree)[pos].predictor);
-      base_bits = EstimateBits(counts.data() + pred * max_symbols,
-                               rounded_counts.data(), max_symbols) +
-                  tot_extra_bits[pred];
+      base_bits =
+          EstimateBits(counts.data() + pred * max_symbols, max_symbols) +
+          tot_extra_bits[pred];
     }
 
     SplitInfo *best = &best_split_nonstatic;
@@ -353,11 +342,9 @@ void FindBestSplit(TreeSamples &tree_samples, float threshold,
               counts_below[sym] += count_increase[i * max_symbols + sym];
               count_increase[i * max_symbols + sym] = 0;
             }
-            float rcost = EstimateBits(counts_above.data(),
-                                       rounded_counts.data(), max_symbols) +
+            float rcost = EstimateBits(counts_above.data(), max_symbols) +
                           tot_extra_bits[pred] - extra_bits_below;
-            float lcost = EstimateBits(counts_below.data(),
-                                       rounded_counts.data(), max_symbols) +
+            float lcost = EstimateBits(counts_below.data(), max_symbols) +
                           extra_bits_below;
             JXL_DASSERT(extra_bits_below <= tot_extra_bits[pred]);
             float penalty = 0;