@jonsneyers 2023-07-12
> Can you give updated speed numbers? Thanks!

I couldn't really measure a big difference after removing that unnecessary "if 0", for jyrki31 the numbers are about the same, for the nonphoto corpus I got 0.046 Mpx/s at e9 instead of 0.045 but I didn't measure accurately enough to consider that an actual difference (just one encode repetition, that already takes a while when doing e9 on a set of images).